refactor(media): Split url resolvers
