+++ packages/editor/src/plugins/media/services/geogebra/index.tsx
@@ -1,4 +1,10 @@
-import { Embed, EmbeddingRenderer, Hosting, ResourceResolver } from '../types'
+import {
+  Embed,
+  EmbeddingRenderer,
+  Hosting,
+  ResourceResolver,
+  URLResolver,
+} from '../types'
 
 export const geogebraResourceResolver: ResourceResolver<
   Hosting.GeoGebra,
@@ -18,4 +24,22 @@ export const GeogebraRenderer: EmbeddingRenderer<Embed.GeoGebraApplet> = ({
       Here would be the embedding for GeoGebra applet <code>{appletId}</code>
     </div>
   )
+}
+
+export const geogebraUrlResolver: URLResolver = {
+  resolvableEmbeddings: [Embed.GeoGebraApplet],
+  resolve(url) {
+    if (url.hostname === 'www.geogebra.org' && url.pathname.startsWith('/m/')) {
+      return {
+        type: 'resourceFound',
+        resource: {
+          hostingService: Hosting.GeoGebra,
+          embeddingType: Embed.GeoGebraApplet,
+          appletId: url.pathname.slice(3),
+        },
+      }
+    } else {
+      return { type: 'cannotResolve' }
+    }
+  },
 }
