+++ packages/editor/src/plugins/media/services/url-resolvers.ts
@@ -1,0 +1,9 @@
+import { geogebraUrlResolver } from './geogebra'
+import { imageUrlResolver } from './html-image'
+import { videoUrlResolver } from './html-video'
+
+export const urlResolvers = [
+  imageUrlResolver,
+  videoUrlResolver,
+  geogebraUrlResolver,
+] as const
